Python_I/IP Geolocalización II/requirements.txt
ipinfo
python-dotenv
folium
requests
ipaddress